CORE-4925 initial
