+++ libs/crypto/crypto-impl/src/test/kotlin/net/corda/crypto/impl/CordaSecurityProviderTests.kt
@@ -1,0 +1,38 @@
+package net.corda.crypto.impl
+
+import net.corda.v5.cipher.suite.CipherSchemeMetadata
+import net.corda.v5.crypto.CompositeKey
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Test
+import org.mockito.kotlin.mock
+
+class CordaSecurityProviderTests {
+    companion object {
+        lateinit var schemeMetadata: CipherSchemeMetadata
+
+        @BeforeAll
+        @JvmStatic
+        fun setup() {
+            schemeMetadata = mock()
+        }
+    }
+
+    @Test
+    fun `get security provider algorithm services`() {
+        val keyFactoryType = "KeyFactory"
+        val signatureType = "Signature"
+        val cordaSecurityProvider = CordaSecurityProvider(schemeMetadata)
+
+        val keyFactoryService = cordaSecurityProvider.getService(keyFactoryType, CompositeKey.KEY_ALGORITHM)
+        val signatureService = cordaSecurityProvider.getService(signatureType, CompositeSignature.SIGNATURE_ALGORITHM)
+
+        assertEquals(keyFactoryType, keyFactoryService!!.type)
+        assertEquals(CompositeKey.KEY_ALGORITHM, keyFactoryService.algorithm)
+        assertEquals(CompositeKeyFactory::class.java.name, keyFactoryService.className)
+
+        assertEquals(signatureType, signatureService!!.type)
+        assertEquals(CompositeSignature.SIGNATURE_ALGORITHM, signatureService.algorithm)
+        assertEquals(CompositeSignature::class.java.name, signatureService.className)
+    }
+}
