+++ settings.gradle
@@ -181,6 +181,7 @@ include 'libs:configuration:configuration-endpoints'
 include 'libs:configuration:configuration-merger'
 include 'libs:configuration:configuration-schema:p2p'
 include 'libs:configuration:configuration-validation'
+include 'libs:crypto:cipher-suite-impl'
 include 'libs:crypto:crypto-core'
 include 'libs:crypto:crypto-flow'
 include 'libs:crypto:crypto-impl'
